attempt to provide map/grep/sort
This fails utterly for so many reasons:

1) Can't override map, grep, or sort at all. Perl's ability to do map
EXPR and map BLOCK makes this absolutely impossible. We can't do the
EXPR versions of these things.

2) Can't return arrays from subs, only lists. This flattens the tied
array, which, because of the way we make our array, only gets the first
item (FETCHSIZE returns 1 and is not re-evaluated). This also completely
removes the tie()...

We may try this again, but instead wrapping around an Iterator::Simple
object maybe?